Code re-factorization for better packages structure
