--- src/main/java/com/unirutas/annotations/PrimaryKey.java
@@ -1,12 +1,0 @@
-package com.unirutas.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface PrimaryKey {
-    String[] value();
-}
