+++ src/main/java/com/unirutas/core/annotations/Column.java
@@ -1,0 +1,17 @@
+package com.unirutas.core.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * An annotation to specify that a field represents a column in a database table.
+ * The value of the annotation is the name of the column.
+ * The name of the column is case-sensitive. It must match the name of the column in the database.
+ */
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Column {
+    String name();
+}
