+++ src/main/java/com/unirutas/repository/AdministrativeRepository.java
@@ -1,0 +1,10 @@
+package com.unirutas.repository;
+
+import com.unirutas.core.database.repository.implementation.sql.SQLGenericRepository;
+import com.unirutas.models.Administrative;
+
+public class AdministrativeRepository extends SQLGenericRepository<Administrative> {
+    public AdministrativeRepository() {
+        super(Administrative.class);
+    }
+}
