--- src/main/java/com/unirutas/repository/implementation/AdministrativeRepository.java
@@ -1,9 +1,0 @@
-package com.unirutas.repository.implementation;
-
-import com.unirutas.models.Administrative;
-
-public class AdministrativeRepository extends GenericRepository<Administrative> {
-    public AdministrativeRepository() {
-        super(Administrative.class);
-    }
-}
