--- src/main/java/com/unirutas/repository/implementation/StudentRepository.java
@@ -1,9 +1,0 @@
-package com.unirutas.repository.implementation;
-
-import com.unirutas.models.Student;
-
-public class StudentRepository extends GenericRepository<Student> {
-    public StudentRepository() {
-        super(Student.class);
-    }
-}
